prepare forms structure
