+++ src/components/v5/common/ActionSidebar/partials/MintTokenForm/index.ts
@@ -1,0 +1,1 @@
+export { default } from './MintTokenForm';
