+++ src/components/v5/common/ActionSidebar/partials/SinglePaymentForm/index.ts
@@ -1,0 +1,1 @@
+export { default } from './SinglePaymentForm';
